refactor(compare): divide compare-divergence.ts em status/navegação
Parte da issue #390 (epic #376, quick-wins de fallow). O arquivo tinha
150 linhas e 8 dependentes; isDocComplete/resolveCompareStatus foram
para compare-assignment-status.ts e findNextPendingDocIndex para
compare-queue-navigation.ts, deixando compare-divergence.ts só com o
núcleo de cálculo (computeDivergentFieldNames). Fan-in de
compare-divergence.ts cai de 8 para 7 dependentes; os dois módulos
novos saem da lista de refactoring targets do fallow. Sem mudança de
comportamento.
