+++ frontend/src/lib/compare-divergence.ts
@@ -103,47 +103,3 @@ export function computeDivergentFieldNames(
 
   return divergent;
 }
-
-export type CompareAssignmentStatus = "pendente" | "em_andamento" | "concluido";
-
-// Pure decision for the reviewer's "comparacao" assignment status, given the
-// canonical divergent fields of the document and the set of fields the reviewer
-// already has a verdict for. `every` is vacuously true for an empty divergent
-// list, so a document with no divergences (e.g. all fused via equivalence)
-// resolves to "concluido" instead of being stuck forever (#217).
-export function resolveCompareStatus(
-  divergentFields: string[],
-  reviewedFields: Set<string>,
-): CompareAssignmentStatus {
-  if (divergentFields.every((fn) => reviewedFields.has(fn))) return "concluido";
-  if (reviewedFields.size === 0) return "pendente";
-  return "em_andamento";
-}
-
-// A document is "complete" in the Compare queue when it has at least one
-// divergent field and every divergent field already has a verdict.
-export function isDocComplete(
-  divergentForDoc: string[] | undefined,
-  reviewsForDoc: Record<string, unknown> | undefined,
-): boolean {
-  if (!divergentForDoc || divergentForDoc.length === 0) return false;
-  if (!reviewsForDoc) return false;
-  return divergentForDoc.every((fn) => !!reviewsForDoc[fn]);
-}
-
-// Index of the next document that still has unreviewed divergences, scanning
-// the queue in its current order and skipping `currentDocId`. Returns -1 when
-// every other document is complete. The server re-sorts the queue on each
-// revalidate (completed docs sink to the bottom), so "next parecer" must be
-// found by completion state — never by `currentIndex + 1`.
-export function findNextPendingDocIndex(
-  docIds: string[],
-  divergentFields: Record<string, string[]>,
-  reviews: Record<string, Record<string, unknown>>,
-  currentDocId: string | undefined,
-): number {
-  return docIds.findIndex(
-    (id) =>
-      id !== currentDocId && !isDocComplete(divergentFields[id], reviews[id]),
-  );
-}
